+++ docs/pub_example/lib/example.g.dart
@@ -1,6 +1,6 @@
 // GENERATED CODE - DO NOT MODIFY BY HAND
 
-part of 'stash.dart';
+part of 'example.dart';
 
 // **************************************************************************
 // SubcommandGenerator
@@ -106,5 +106,69 @@ class ApplyCommand extends Command {
     final results = argResults!;
     return userMethod(
         stashRef: results['stash-ref'] != null ? results['stash-ref'] : '0');
+  }
+}
+
+// **************************************************************************
+// CliRunnerGenerator
+// **************************************************************************
+
+class _$GitRunner extends CommandRunner {
+  _$GitRunner()
+      : super(
+          'git',
+          'A command-line interface for version control.',
+        ) {
+    final upcastedType = (this as GitRunner);
+    addCommand(MergeCommand(upcastedType.merge));
+  }
+}
+
+class MergeCommand extends Command {
+  MergeCommand(this.userMethod);
+
+  final Function({
+    required String branch,
+    MergeStrategy strategy,
+    bool? commit,
+  }) userMethod;
+
+  @override
+  String get name => 'merge';
+
+  @override
+  String get description => 'Join two or more development histories together.';
+
+  @override
+  ArgParser get argParser => ArgParser()
+    ..addOption(
+      'branch',
+      mandatory: true,
+    )
+    ..addOption(
+      'strategy',
+      defaultsTo: 'ort',
+      mandatory: false,
+      allowed: [
+        'ort',
+        'recursive',
+        'resolve',
+        'octopus',
+        'ours',
+        'subtree',
+      ],
+    )
+    ..addFlag('commit');
+
+  @override
+  Future<void> run() {
+    final results = argResults!;
+    return userMethod(
+      branch: results['branch'],
+      strategy: results['strategy'] != null
+          ? EnumParser(MergeStrategy.values).parse(results['strategy'])
+          : MergeStrategy.ort,
+      commit: results['commit'] != null ? results['commit'] : null,
+    );
   }
 }