+++ src/cli_annotations/lib/src/args_annotations.dart
@@ -1,13 +1,13 @@
-// TODO: the following two annotations are only necessary if macros doesn't
-// allow introspection of doc comments and/or default values of parameters/fields.
-class Comment {
-  final String comment;
+// // TODO: the following two annotations are only necessary if macros doesn't
+// // allow introspection of doc comments and/or default values of parameters/fields.
+// class Comment {
+//   final String comment;
 
-  const Comment(this.comment);
-}
+//   const Comment(this.comment);
+// }
 
-class Default {
-  final String value;
+// class Default {
+//   final String value;
 
-  const Default(this.value);
-}
+//   const Default(this.value);
+// }