+++ src/cli_annotations/test/src/args_test.dart
@@ -1,3 +1,5 @@
+// ignore_for_file: unused_local_variable
+
 import 'package:cli_annotations/cli_annotations.dart';
 import 'package:test/test.dart';
 
@@ -9,5 +11,17 @@ void main() {
       category: 'test category',
     );
     cliCommand.toString();
+
+    final cliRunner = CliRunner(
+      name: 'test',
+      description: 'test description',
+    );
+
+    final toString = cliRunner.toString();
+
+    final cliSubcommand = CliSubcommand(
+      name: 'test',
+      description: 'test description',
+    );
   });
 }