+++ src/cli_gen/CHANGELOG.md
@@ -1,3 +1,7 @@
+## 0.1.0-dev.3
+
+- docs: Added example.
+
 ## 0.1.0-dev.2
 
 - fix: Images in README.md should now be displayed correctly.